test: restructure suite into unit, integration, e2e, and vite projects
Split src/KCC.Web.Tests into tests/KCC.UnitTests and add integration, e2e, and vite test projects. Migrate from xunit to TUnit on the Microsoft Testing Platform and wire up the solution, package versions, and configs.
