+++ KitchenCommandCenter.sln
@@ -1,4 +1,5 @@
-﻿Microsoft Visual Studio Solution File, Format Version 12.00
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 18
 VisualStudioVersion = 18.0.11222.15
 MinimumVisualStudioVersion = 10.0.40219.1
@@ -8,9 +9,15 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KCC.Web", "src\KCC.Web\KCC.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KCC.Admin", "src\KCC.Admin\KCC.Admin.csproj", "{2D2FFEA2-3E3B-48AD-AC43-6DD7A4362768}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KCC.Web.Tests", "src\KCC.Web.Tests\KCC.Web.Tests.csproj", "{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KCC.ResourceStrings", "src\KCC.ResourceStrings\KCC.ResourceStrings.csproj", "{C4BEF8BF-ED94-4F75-BF07-61F76EAA0C67}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{0AB3BF05-4346-4AA6-1389-037BE0695223}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KCC.UnitTests", "tests\KCC.UnitTests\KCC.UnitTests.csproj", "{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KCC.IntegrationTests", "tests\KCC.IntegrationTests\KCC.IntegrationTests.csproj", "{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KCC.E2ETests", "tests\KCC.E2ETests\KCC.E2ETests.csproj", "{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -46,18 +53,6 @@ Global
 		{2D2FFEA2-3E3B-48AD-AC43-6DD7A4362768}.Release|x64.Build.0 = Release|Any CPU
 		{2D2FFEA2-3E3B-48AD-AC43-6DD7A4362768}.Release|x86.ActiveCfg = Release|Any CPU
 		{2D2FFEA2-3E3B-48AD-AC43-6DD7A4362768}.Release|x86.Build.0 = Release|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Debug|x64.Build.0 = Debug|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Debug|x86.Build.0 = Debug|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Release|x64.ActiveCfg = Release|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Release|x64.Build.0 = Release|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Release|x86.ActiveCfg = Release|Any CPU
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F}.Release|x86.Build.0 = Release|Any CPU
 		{C4BEF8BF-ED94-4F75-BF07-61F76EAA0C67}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C4BEF8BF-ED94-4F75-BF07-61F76EAA0C67}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C4BEF8BF-ED94-4F75-BF07-61F76EAA0C67}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -70,6 +65,42 @@ Global
 		{C4BEF8BF-ED94-4F75-BF07-61F76EAA0C67}.Release|x64.Build.0 = Release|Any CPU
 		{C4BEF8BF-ED94-4F75-BF07-61F76EAA0C67}.Release|x86.ActiveCfg = Release|Any CPU
 		{C4BEF8BF-ED94-4F75-BF07-61F76EAA0C67}.Release|x86.Build.0 = Release|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Debug|x64.Build.0 = Debug|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Debug|x86.Build.0 = Debug|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Release|x64.ActiveCfg = Release|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Release|x64.Build.0 = Release|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Release|x86.ActiveCfg = Release|Any CPU
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79}.Release|x86.Build.0 = Release|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Debug|x64.Build.0 = Debug|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Debug|x86.Build.0 = Debug|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Release|x64.ActiveCfg = Release|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Release|x64.Build.0 = Release|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Release|x86.ActiveCfg = Release|Any CPU
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF}.Release|x86.Build.0 = Release|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Debug|x64.Build.0 = Debug|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Debug|x86.Build.0 = Debug|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Release|x64.ActiveCfg = Release|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Release|x64.Build.0 = Release|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Release|x86.ActiveCfg = Release|Any CPU
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -77,8 +108,10 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{1ECDCC7B-F704-425A-B777-8B1CADB97895} = {56A60E2B-E452-4E2C-A1C3-57418060BF44}
 		{2D2FFEA2-3E3B-48AD-AC43-6DD7A4362768} = {56A60E2B-E452-4E2C-A1C3-57418060BF44}
-		{2350BAA1-0335-4D36-B19E-8331B1AEAD7F} = {56A60E2B-E452-4E2C-A1C3-57418060BF44}
 		{C4BEF8BF-ED94-4F75-BF07-61F76EAA0C67} = {56A60E2B-E452-4E2C-A1C3-57418060BF44}
+		{6CC352FD-AC5C-4C62-B1ED-0DCC9F11AF79} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{AD607C9A-DE02-45AA-92E6-CA57C50D07BF} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{B2F55766-83FE-44CC-A23D-C93F90CBE5F5} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {5C84EBAB-A5CB-4932-A723-87637A16ADFC}
