+++ src/KCC.Web/KCC.Web.csproj
@@ -53,7 +53,8 @@
     </ItemGroup>
 
     <ItemGroup>
-        <InternalsVisibleTo Include="KCC.Web.Tests" />
+        <InternalsVisibleTo Include="KCC.UnitTests" />
+        <InternalsVisibleTo Include="KCC.IntegrationTests" />
     </ItemGroup>
 
 </Project>
