+++ src/KCC.Web/eslint.config.ts
@@ -15,7 +15,7 @@ export default defineConfigWithVueTs(
 
   {
     name: 'app/files-to-ignore',
-    ignores: ['**/dist/**', '**/dist-ssr/**', '**/coverage/**', '**/wwwroot/**'],
+    ignores: ['**/dist/**', '**/wwwroot/**', '**/wwwroot/**'],
   },
 
   prettierConfig,
