+++ tests/KCC.E2ETests/Config/Constants.cs
@@ -1,0 +1,6 @@
+namespace KCC.E2ETests.Config;
+
+public static class Constants
+{
+    public const string TESTING_DOMAIN = "http://localhost:58671";
+}
