+++ tests/KCC.IntegrationTests/appsettings.CI.json
@@ -1,0 +1,5 @@
+{
+  "ConnectionStrings": {
+    "CMSConnectionString": "Data Source=localhost;Initial Catalog=KitchenCommandCenter;Integrated Security=False;User ID=sa;Password=Pass@12345;Persist Security Info=False;Connect Timeout=60;Encrypt=False;Current Language=English;"
+  }
+}
