+++ tests/KCC.IntegrationTests/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "ConnectionStrings": {
+    "CMSConnectionString": ""
+  }
+}
