+++ tests/KCC.UnitTests/KCC.UnitTests.csproj
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <TargetFramework>net10.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <IsPackable>false</IsPackable>
+        <IsTestProject>true</IsTestProject>
+        <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
+        <WarningsNotAsErrors>NU1608;NU1901;NU1902;NU1903</WarningsNotAsErrors>
+    </PropertyGroup>
+    <ItemGroup>
+        <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="Microsoft.NET.Test.Sdk" />
+        <PackageReference Include="Moq" />
+        <PackageReference Include="TUnit" />
+    </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\..\src\KCC.Web\KCC.Web.csproj" />
+        <ProjectReference Include="..\..\src\KCC.Admin\KCC.Admin.csproj" />
+        <ProjectReference Include="..\..\src\KCC.ResourceStrings\KCC.ResourceStrings.csproj" />
+    </ItemGroup>
+</Project>
